1) Movies try..catch from the ViewModel to the datalayer 2) Movies the NetworkResponseState to the domain layer to be accessed by presentation and data 3) Set the mutable state to have an initial null state that will represent IDLE 4) Moves the idle object from the NetworkResponseState
