+++ domain/src/main/java/me/androidbox/domain/authentication/usecase/LoginUseCase.kt
@@ -1,7 +1,8 @@
 package me.androidbox.domain.authentication.usecase
 
+import me.androidbox.domain.authentication.NetworkResponseState
 import me.androidbox.domain.authentication.model.LoginModel
 
 fun interface LoginUseCase {
-    suspend fun execute(email: String, password: String): LoginModel
+    suspend fun execute(email: String, password: String): NetworkResponseState<LoginModel>
 }
