+++ domain/src/main/java/me/androidbox/domain/authentication/usecase/RegisterUseCase.kt
@@ -1,5 +1,7 @@
 package me.androidbox.domain.authentication.usecase
 
+import me.androidbox.domain.authentication.NetworkResponseState
+
 fun interface RegisterUseCase {
-    suspend fun execute(fullName: String, email: String, password: String)
+    suspend fun execute(fullName: String, email: String, password: String): NetworkResponseState<Unit>
 }
