+++ domain/src/main/java/me/androidbox/domain/authentication/usecase/imp/LoginUseCaseImp.kt
@@ -1,12 +1,13 @@
 package me.androidbox.domain.authentication.usecase.imp
 
+import me.androidbox.domain.authentication.NetworkResponseState
 import me.androidbox.domain.authentication.model.LoginModel
 import me.androidbox.domain.authentication.remote.AuthenticationRepository
 import me.androidbox.domain.authentication.usecase.LoginUseCase
 import javax.inject.Inject
 
 class LoginUseCaseImp @Inject constructor(private val authenticationRepository: AuthenticationRepository): LoginUseCase {
-    override suspend fun execute(email: String, password: String): LoginModel {
+    override suspend fun execute(email: String, password: String): NetworkResponseState<LoginModel> {
         return authenticationRepository.loginUser(email, password)
     }
 }
