+++ domain/src/main/java/me/androidbox/domain/authentication/usecase/imp/RegisterUseCaseImp.kt
@@ -1,5 +1,6 @@
 package me.androidbox.domain.authentication.usecase.imp
 
+import me.androidbox.domain.authentication.NetworkResponseState
 import me.androidbox.domain.authentication.remote.AuthenticationRepository
 import me.androidbox.domain.authentication.usecase.RegisterUseCase
 import javax.inject.Inject
@@ -8,8 +9,8 @@ class RegisterUseCaseImp @Inject constructor(
     private val authenticationRepository: AuthenticationRepository
 ) : RegisterUseCase {
 
-    override suspend fun execute(fullName: String, email: String, password: String) {
-        authenticationRepository.registerUser(
+    override suspend fun execute(fullName: String, email: String, password: String): NetworkResponseState<Unit> {
+        return authenticationRepository.registerUser(
             fullName = fullName,
             email = email,
             password = password)
