+++ presentation/src/main/java/me/androidbox/presentation/HomeActivity.kt
@@ -5,6 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import dagger.hilt.android.AndroidEntryPoint
 import me.androidbox.presentation.login.screen.LoginScreen
+import me.androidbox.presentation.login.screen.RegisterScreen
 import me.androidbox.presentation.ui.theme.BusbyTaskyTheme
 
 @AndroidEntryPoint
@@ -13,7 +14,7 @@ class HomeActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             BusbyTaskyTheme {
-                LoginScreen {
+                LoginScreen() {
 
                 }
             }
